Update `docker/cli` with upstream changes
This is still depending on a fork, but with way less differents
commits (several changes have been already upstreamed) and with an
up-to-date `docker/cli`.

One big change to notice is the `enabled` becomes
`x-enabled` (using *extras* field of the composefile format), and thus
does not require any change to the composefile struct.

Signed-off-by: Vincent Demeester <vincent@sbr.pm>