Fix helm generation
This bump the docker/cli version to get the fix required on
`ExtractVariables`. It now takes a pattern other than the default,
which is something we require on `docker/app` as the pattern we
use (when interpolating) is a bit different from the default one.

It also adds a small message on the e2e test errors to see which file
was wrong.

Signed-off-by: Vincent Demeester <vincent@sbr.pm>