Build invocation image locally
Since we are using `docker context` that is not yet released we need to
cross-compile the master branch of the cli on Jenkins for our e2e tests
to run.
This takes a lot of time when not on Linux and it doesn't make sense to
do locally.

With the new Dockerfile.invocation-image it's possible to run
`make invocation-image` and build the invocation image without
recompiling the cli.

Signed-off-by: Djordje Lukic <djordje.lukic@docker.com>