Fix coverage on Jenkins
We are running multiple stages in parallel that need the same invocation
image built in a previous step. Because those stages can run on the same
Jenkins slave we need a way to uniquely identify and tag the invocation
image so that we can safely cleanup after the stage is finshed. Docker
save/load doesn't allow the tag to be changed so we need to work around
this and tag manually the image that is built.

Having unique tag names for all stages ensures that we won't try to use
or remove an invocation image that doesn't belong to that stage,
effectively isolating all the stages that run in parallel.

Signed-off-by: Djordje Lukic <djordje.lukic@docker.com>