Drop invalid TODO regarding use of `command.Cli`
The function `newRootCommand` needs to call the `Initialize` method which is
part of `*command.DockerCli` but is not included in the `command.Cli`
interface (since it is effectively internal infra).

When the `dockerCli` is passed on to `addCommands` it is a `command.Cli` from
that point on, which is what we really wanted.

Signed-off-by: Ian Campbell <ijc@docker.com>