resto: Include URI for failed pushes
This allows the error message for a failed push to be more informative
by including the URI of the failed push. For example, the code
currently doesn't honor the "namespace" attribute in
hello-world.dockerapp, which lead to some confusion on my part. With
this change the error became:

 ./bin/docker-app-linux -D push hello-world.dockerapp
  Error: Unable to push to: https://registry-1.docker.io/library/hello-world
  errors:
  denied: requested access to the resource is denied
  unauthorized: authentication required

Not the greatest error, but it might save someone the effort of
having to recompile with printfs to understand.

Signed-off-by: Andy Doan <andy@foundries.io>