Push prints error messages correctly
When using docker-app as a plugin, error messages did not appear in the
console.
This is due to the fact that we did never un-mute the docker CLI structure
when Pushing, and that the RunPlugin uses dockerCli.Err() to report
error messages to the user.

This fixes the issue by using the `defer muteDockerCli(cli)()` pattern
that is used everywhere else.

Signed-off-by: Simon Ferquel <simon.ferquel@docker.com>