Define constants for all custom actions
There is one small semantic change here which is that the `prepareCustomAction`
helper now takes the full qualified/namespaced name and not just the suffix.

This is no longer valid to be used as the argument to `claim.New(...)` however
as far as I can tell that string is completely arbitrary and is never used (we
never store this claim) so it might as well be some hardcoded string.

Signed-off-by: Ian Campbell <ijc@docker.com>