ci: run e2e tests via gotestsum and record junit result
As well as gotestsum we now also need to build `test2json` because test
binaries build with `go test -c` do not support the `-test` option. That option
just wraps the test run in the `test2json` code which translates from the
standard output syntax to json. (See https://github.com/golang/go/issues/31051
for more info). This uses `cmd/test2json` the source of which is part of the go
distribution and is therefore available in the base images.

Once we have them then we "simply" chain them together, `gotestsum` takes a
`--raw-command` for which we pass `test2json`. In turn `test2json` takes a
command which it will wrap and we pass the actual e2e test binary.

Running `test2json` manually gives us the opportunity to set the package name
(with `-p`) we give it a distinct name rather than munging the output after the
fact like we do with the unit tests.

Since the only references to these tools are in `Jenkinsfile.*` we don't need
to worry about making them available for non-containerised runs.

`test2json` requires verbose output, so add `-test.v` in one case.

Signed-off-by: Ian Campbell <ijc@docker.com>