ci: adjust testsuite names in junit
For unit tests handle the top-level specially so that second-level tests do not
have a leading `/` which confuses things it seems, by creating an entry in the
table with an empty label.

For e2e tests there are no subtests, so do a simpler transformation.

Arrange all the tests in a somewhat hierarchical way, which after some
experimentation seems to lead to a satisfactory rendering in the Jenkins UI.
Other things which were tried lead to splitting tests with some things at the
top-level and others under the `(root)`.

Signed-off-by: Ian Campbell <ijc@docker.com>