ci: ensure we always get the test results even if they fail
The previous strucuture meant that a failure skipped the result extraction.

I considered switching to using `-v` to mount the output dirs to the host, but
this would result in root owned files. Instead I split the `Makefile` target
into two parts which the CI can call independently, the original target remains
with the same overall behaviour for end user use.

Signed-off-by: Ian Campbell <ijc@docker.com>