inspect: sort services, networks, volumes and secrets by name
Networks, Volumes and Secrets are maps so iteration is randomized by the
runtime. Although Services are an slice here they have, at some point, be
laundered through a map and so also need sorting.

Test by adding a second instance of each to the "full" test case and by always
inspecting twice (in case the first one got lucky).

Fixes: #525

Signed-off-by: Ian Campbell <ijc@docker.com>