Remove `--insecure-registries` flag
The `--insecure-registries` flag is removed to have a consistent UX with
Docker CLI.

Instead of this flag, the list of insecure registries is retrieved from
the engine. See https://docs.docker.com/registry/insecure/ to configure
your engine to allow communication to an insecure registry.

See the engine API documentation at
https://docs.docker.com/engine/api/v1.40/#operation/SystemInfo
and `RegistryConfig.IndexConfigs` fields.

Signed-off-by: Yves Brissaud <yves.brissaud@docker.com>