split image ls output to repo and tag
Split the app image ls output up so that the REPOSITORY and TAG are
separate columns, the same as the standard docker image ls.

The APP IMAGE ID column is now a default column, same as the ID column
in docker image ls.

As this ID is randomly generated the e2e tests have
been updated to use regexes when asserting the command output.

Signed-off-by: Nick Adcock <nick.adcock@docker.com>