Added created date to app images
Added a created date field to the custom payload of the bundle.json.
This is packaged as part of the app image at build time.

The CREATED column has been added to the `app image ls` output. This
displays the time since the app image was built in the same human
readable format as `image ls`.

As the creation date is dynamic the app image ID will be unique for each
build even if nothing else has changed. To adjust the tests for this the
golden bundle.json is now checked using regex matching instead of string
matching.

Simalarly, the e2e tests for `app image ls` are checked with regexes for
the CREATED column as the exact value will depend on how long the tests
take to run.

Signed-off-by: Nick Adcock <nick.adcock@docker.com>