use relocation map on fixup
When the fixup is performed, if the image cannot be resolved
try with the image in the relocation map if it exists.

This way a pulled application can be pushed without to require
to download locally and push all the images are they already exist
in the registry

See https://github.com/docker/cnab-to-oci/pull/82 to have
more details about the change.

Signed-off-by: Yves Brissaud <yves.brissaud@docker.com>